Split 2-OPT into ReverseSegment and SwapTails
