+++ configs/vrptw.toml
@@ -9,13 +9,14 @@ node_ops = [
     "Exchange32",
     "Exchange33",
     "MoveTwoClientsReversed",
-    "TwoOpt",
+    "ReverseSegment",
 ]
 
 
 route_ops = [
     "RelocateStar",
     "SwapStar",
+    "SwapTails",
 ]
