+++ pyvrp/cpp/search/SwapTails.cpp
@@ -1,60 +1,14 @@
-#include "TwoOpt.h"
+#include "SwapTails.h"
 
 #include "Route.h"
 
 #include <cassert>
 
-using pyvrp::Cost;
-using pyvrp::search::TwoOpt;
+using pyvrp::search::SwapTails;
 
-Cost TwoOpt::evalWithinRoute(Route::Node *U,
-                             Route::Node *V,
-                             CostEvaluator const &costEvaluator) const
-{
-    assert(U->route() == V->route());
-    auto *route = U->route();
-
-    Cost deltaCost
-        = -static_cast<Cost>(route->distance())
-          - costEvaluator.loadPenalty(route->load(), route->capacity())
-          - costEvaluator.twPenalty(route->timeWarp());
-
-    // Current situation is U -> n(U) -> ... -> V -> n(V). Proposed move is
-    // U -> V -> p(V) -> ... -> n(U) -> n(V). This reverses the segment from
-    // n(U) to V.
-    DistanceSegment dist = route->before(U->idx());
-    for (size_t idx = V->idx(); idx != U->idx(); --idx)
-        dist = DistanceSegment::merge(
-            data.distanceMatrix(), dist, route->at(idx));
-    dist = DistanceSegment::merge(
-        data.distanceMatrix(), dist, route->after(V->idx() + 1));
-
-    deltaCost += static_cast<Cost>(dist.distance());
-
-    if (deltaCost >= 0)
-        return deltaCost;
-
-    LoadSegment ls = route->before(U->idx());
-    for (size_t idx = V->idx(); idx != U->idx(); --idx)
-        ls = LoadSegment::merge(ls, route->at(idx));
-    ls = LoadSegment::merge(ls, route->after(V->idx() + 1));
-
-    deltaCost += costEvaluator.loadPenalty(ls.load(), route->capacity());
-
-    DurationSegment ds = route->before(U->idx());
-    for (size_t idx = V->idx(); idx != U->idx(); --idx)
-        ds = DurationSegment::merge(data.durationMatrix(), ds, route->at(idx));
-    ds = DurationSegment::merge(
-        data.durationMatrix(), ds, route->after(V->idx() + 1));
-
-    deltaCost += costEvaluator.twPenalty(ds.timeWarp(route->maxDuration()));
-
-    return deltaCost;
-}
-
-Cost TwoOpt::evalBetweenRoutes(Route::Node *U,
-                               Route::Node *V,
-                               CostEvaluator const &costEvaluator) const
+pyvrp::Cost SwapTails::evaluate(Route::Node *U,
+                                Route::Node *V,
+                                CostEvaluator const &costEvaluator)
 {
     auto const *uRoute = U->route();
     auto const *vRoute = V->route();
@@ -183,66 +137,49 @@ Cost TwoOpt::evalBetweenRoutes(Route::Node *U,
     return deltaCost;
 }
 
-void TwoOpt::applyWithinRoute(Route::Node *U, Route::Node *V) const
+pyvrp::Cost
+SwapTails::evaluate(Route *U, Route *V, CostEvaluator const &costEvaluator)
 {
-    auto *nU = n(U);
+    move = {};
 
-    while (V->idx() > nU->idx())
-    {
-        auto *pV = p(V);
-        Route::swap(nU, V);
-        nU = n(V);  // after swap, V is now nU
-        V = pV;
-    }
+    for (size_t uIdx = 0; uIdx != U->size() + 1; ++uIdx)
+        for (size_t vIdx = 0; vIdx != V->size() + 1; ++vIdx)
+        {
+            auto *nodeU = (*U)[uIdx];
+            auto *nodeV = (*V)[vIdx];
+            auto deltaCost = evaluate(nodeU, nodeV, costEvaluator);
+
+            if (deltaCost < move.deltaCost)
+            {
+                move.deltaCost = deltaCost;
+                move.U = nodeU;
+                move.V = nodeV;
+            }
+        }
+
+    return move.deltaCost;
 }
 
-void TwoOpt::applyBetweenRoutes(Route::Node *U, Route::Node *V) const
+void SwapTails::apply(Route *U, Route *V) const
 {
-    auto *nU = n(U);
-    auto *nV = n(V);
+    auto *nU = n(move.U);
+    auto *nV = n(move.V);
 
-    auto insertIdx = U->idx() + 1;
+    auto insertIdx = move.U->idx() + 1;
     while (!nV->isDepot())
     {
         auto *node = nV;
         nV = n(nV);
-        V->route()->remove(node->idx());
-        U->route()->insert(insertIdx++, node);
+        V->remove(node->idx());
+        U->insert(insertIdx++, node);
     }
 
-    insertIdx = V->idx() + 1;
+    insertIdx = move.V->idx() + 1;
     while (!nU->isDepot())
     {
         auto *node = nU;
         nU = n(nU);
-        U->route()->remove(node->idx());
-        V->route()->insert(insertIdx++, node);
+        U->remove(node->idx());
+        V->insert(insertIdx++, node);
     }
-}
-
-Cost TwoOpt::evaluate(Route::Node *U,
-                      Route::Node *V,
-                      CostEvaluator const &costEvaluator)
-{
-    assert(U->route());
-    assert(V->route());
-
-    if (U->route()->idx() > V->route()->idx())  // tackled in a later iteration
-        return 0;
-
-    if (U->route() != V->route())
-        return evalBetweenRoutes(U, V, costEvaluator);
-
-    if (U->idx() + 1 >= V->idx())  // tackled in a later iteration
-        return 0;
-
-    return evalWithinRoute(U, V, costEvaluator);
-}
-
-void TwoOpt::apply(Route::Node *U, Route::Node *V) const
-{
-    if (U->route() == V->route())
-        applyWithinRoute(U, V);
-    else
-        applyBetweenRoutes(U, V);
 }
