+++ pyvrp/search/_search.pyi
@@ -35,9 +35,9 @@ class Exchange32(NodeOperator): ...
 class Exchange33(NodeOperator): ...
 class MoveTwoClientsReversed(NodeOperator): ...
 class RelocateStar(RouteOperator): ...
-class SwapRoutes(RouteOperator): ...
+class ReverseSegment(NodeOperator): ...
 class SwapStar(RouteOperator): ...
-class TwoOpt(NodeOperator): ...
+class SwapTails(RouteOperator): ...
 
 class LocalSearch:
     def __init__(
