+++ tests/search/test_SwapTails.py
@@ -1,0 +1,280 @@
+import numpy as np
+import pytest
+from numpy.testing import assert_, assert_allclose, assert_equal
+
+from pyvrp import (
+    Client,
+    CostEvaluator,
+    Depot,
+    ProblemData,
+    RandomNumberGenerator,
+    Solution,
+    VehicleType,
+)
+from pyvrp import Route as SolRoute
+from pyvrp.search import LocalSearch, SwapTails
+from pyvrp.search._search import Node, Route
+
+
+def test_evaluate_empty_routes(ok_small):
+    """
+    Tests that evaluate() returns 0 when one or both of the routes are empty.
+    """
+    data = ok_small.replace(
+        vehicle_types=[
+            VehicleType(3, capacity=10),
+            VehicleType(3, capacity=10),
+        ]
+    )
+
+    route1 = Route(data, idx=0, vehicle_type=0)
+    route2 = Route(data, idx=1, vehicle_type=1)
+    route3 = Route(data, idx=2, vehicle_type=0)
+
+    route1.append(Node(loc=1))
+
+    route1.update()
+    route2.update()
+
+    op = SwapTails(data)
+    cost_eval = CostEvaluator(1, 1)
+
+    # Vehicle types are no longer the same, but one of the routes is empty.
+    # That situation is not currently handled.
+    assert_(route1.vehicle_type != route2.vehicle_type)
+    assert_allclose(op.evaluate(route1, route2, cost_eval), 0)
+    assert_allclose(op.evaluate(route2, route1, cost_eval), 0)
+
+    # Both routes are empty, but of different vehicle type as well.
+    assert_equal(len(route2), len(route3))
+    assert_allclose(op.evaluate(route3, route2, cost_eval), 0)
+
+
+def test_evaluate_shift_time_window_differences(ok_small):
+    """
+    Tests that SwapTails correctly evaluates changes in time warp due to
+    different shift time windows.
+    """
+    data = ok_small.replace(
+        vehicle_types=[
+            VehicleType(capacity=10, tw_early=10_000, tw_late=15_000),
+            VehicleType(capacity=10, tw_early=15_000, tw_late=20_000),
+        ]
+    )
+
+    route1 = Route(data, idx=0, vehicle_type=0)
+    for loc in [1, 4]:  # depot -> 1 -> 4 -> depot
+        route1.append(Node(loc=loc))
+    route1.update()
+
+    route2 = Route(data, idx=1, vehicle_type=1)
+    for loc in [3, 2]:  # depot -> 3 -> 2 -> depot
+        route2.append(Node(loc=loc))
+    route2.update()
+
+    # Without shift time windows, both routes are feasible, and there is slack
+    # on either route: the first route can start between [14'056, 16'003], and
+    # the second between [9'002, 13'369]. Neither route can complete its visits
+    # within the shift time window of its assigned vehicle type. However, the
+    # other type has a shift duration that is much better aligned with its
+    # route. Thus, we should have that swapping the vehicle types results in
+    # a lower cost, due to decreased time warp on the routes.
+    op = SwapTails(data)
+    cost_eval = CostEvaluator(1, 1)
+    assert_(op.evaluate(route1, route2, cost_eval) < 0)
+
+
+def test_evaluate_max_duration_constraints(ok_small):
+    """
+    Tests that SwapTails correctly evaluates changes in time warp due to
+    different maximum duration constraints.
+    """
+    data = ok_small.replace(
+        vehicle_types=[
+            VehicleType(capacity=10, max_duration=3_000),
+            VehicleType(capacity=10),
+        ]
+    )
+
+    route1 = Route(data, idx=0, vehicle_type=0)
+    for loc in [1, 4]:  # depot -> 1 -> 4 -> depot
+        route1.append(Node(loc=loc))
+    route1.update()
+
+    route2 = Route(data, idx=1, vehicle_type=1)
+    for loc in [3, 2]:  # depot -> 3 -> 2 -> depot
+        route2.append(Node(loc=loc))
+    route2.update()
+
+    # First route takes 5'332, which is 2'332 more than its maximum duration
+    # allows. There is no other source of time warp, so the total route time
+    # warp must be 2'332.
+    assert_allclose(route1.duration(), 5_332)
+    assert_allclose(route1.time_warp(), 2_332)
+
+    # Second route takes 5'323, and has no maximum duration constraint. There
+    # is no other source of time warp, so total route time warp must be zero.
+    assert_allclose(route2.duration(), 5_323)
+    assert_allclose(route2.time_warp(), 0)
+
+    # Appending route1 to route2 avoids the maximum route duration constraint
+    # on route1 completely, which is much better than the current solution
+    # because it resolves all time warp:
+    #  * new distance: 7621
+    #  * new load: 18 (8 excess)
+    #  * new time warp: 0
+    op = SwapTails(data)
+    cost_eval = CostEvaluator(1, 1)
+    assert_allclose(
+        op.evaluate(route1, route2, cost_eval),
+        7621 + 8 - route1.distance() - route2.distance() - route1.time_warp(),
+    )
+
+    op.apply(route1, route2)
+    route1.update()
+    route2.update()
+
+    assert_equal(len(route1), 0)
+    assert_equal(len(route2), 4)
+
+    assert_allclose(route2.distance(), 7621)
+    assert_allclose(route2.time_warp(), 0)
+    assert_allclose(route2.excess_load(), 8)
+
+
+def test_evaluate_with_different_depots():  #
+    """
+    Tests that SwapTails correctly evaluates distance changes due to different
+    start and end depots of different vehicle types.
+    """
+    data = ProblemData(
+        clients=[Client(x=1, y=1), Client(x=4, y=4)],
+        depots=[Depot(x=0, y=0), Depot(x=5, y=5)],
+        vehicle_types=[VehicleType(depot=0), VehicleType(depot=1)],
+        distance_matrix=[
+            [0, 10, 2, 8],
+            [10, 0, 8, 2],
+            [2, 8, 0, 6],
+            [8, 2, 6, 0],
+        ],
+        duration_matrix=np.zeros((4, 4), dtype=int),
+    )
+
+    # First route is first depot -> second client -> first depot.
+    route1 = Route(data, idx=0, vehicle_type=0)
+    route1.append(Node(loc=3))
+    route1.update()
+
+    # Second route is second depot -> first client -> second depot.
+    route2 = Route(data, idx=1, vehicle_type=1)
+    route2.append(Node(loc=2))
+    route2.update()
+
+    op = SwapTails(data)
+    cost_eval = CostEvaluator(1, 1)
+
+    # The routes each cost 16 distance which is not as efficient as swapping
+    # them, as that would reduce each route's cost to 4, for an improvement
+    # of 2 * 12 = 24.
+    assert_allclose(route1.distance(), 16)
+    assert_allclose(route2.distance(), 16)
+    assert_allclose(op.evaluate(route1, route2, cost_eval), -24)
+
+
+@pytest.mark.parametrize(
+    "vehicle_types",
+    [
+        [VehicleType(capacity=10), VehicleType(capacity=10)],
+        [VehicleType(capacity=10), VehicleType(capacity=8)],
+        [VehicleType(capacity=9), VehicleType(capacity=9)],
+        [VehicleType(capacity=8), VehicleType(capacity=8)],
+    ],
+)
+def test_OkSmall_multiple_vehicle_types(
+    ok_small, vehicle_types: list[VehicleType]
+):
+    """
+    This test evaluates a move that is improving for some of the vehicle types,
+    and not for others. Because the granular neighbourhood is very constrained,
+    only a single move can be applied. When it's better to do so, it should
+    have been applied, and when it's not better, it should not.
+    """
+    data = ok_small.replace(vehicle_types=vehicle_types)
+
+    cost_evaluator = CostEvaluator(10_000, 6)  # large capacity penalty
+    rng = RandomNumberGenerator(seed=42)
+
+    neighbours: list[list[int]] = [[], [2], [], [], []]  # only 1 -> 2
+    ls = LocalSearch(data, rng, neighbours)
+    ls.add_route_operator(SwapTails(data))
+
+    routes1 = [SolRoute(data, [1, 3], 0), SolRoute(data, [2, 4], 1)]
+    sol1 = Solution(data, routes1)
+
+    routes2 = [SolRoute(data, [1, 4], 0), SolRoute(data, [2, 3], 1)]
+    sol2 = Solution(data, routes2)
+
+    cost1 = cost_evaluator.penalised_cost(sol1)
+    cost2 = cost_evaluator.penalised_cost(sol2)
+    assert_(not np.allclose(cost1, cost2))
+
+    # Using the local search, the result should not get worse.
+    improved_sol1 = ls.intensify(sol1, cost_evaluator, overlap_tolerance=1)
+    improved_sol2 = ls.intensify(sol2, cost_evaluator, overlap_tolerance=1)
+
+    expected_sol = sol1 if cost1 < cost2 else sol2
+    assert_equal(improved_sol1, expected_sol)
+    assert_equal(improved_sol2, expected_sol)
+
+
+def test_move_involving_empty_routes():
+    """
+    This test checks that a SwapTails move that turns an empty route into a
+    non-empty route (or vice-versa) is correctly evaluated.
+    """
+    data = ProblemData(
+        clients=[Client(x=1, y=1), Client(x=1, y=0)],
+        depots=[Depot(x=0, y=0)],
+        vehicle_types=[
+            VehicleType(fixed_cost=10),
+            VehicleType(fixed_cost=100),
+        ],
+        distance_matrix=np.zeros((3, 3), dtype=int),
+        duration_matrix=np.zeros((3, 3), dtype=int),
+    )
+
+    route1 = Route(data, idx=0, vehicle_type=0)
+    route2 = Route(data, idx=1, vehicle_type=1)
+
+    for loc in [1, 2]:
+        route1.append(Node(loc=loc))
+
+    route1.update()  # depot -> 1 -> 2 -> depot
+    route2.update()  # depot -> depot
+
+    op = SwapTails(data)
+    cost_eval = CostEvaluator(0, 0)
+
+    # route1 is non-empty, route2 is empty. The only difference between the
+    # routes is that the vehicle servicing route2 is more expensive than the
+    # one servicing route1. There is thus no improving move here.
+    assert_allclose(op.evaluate(route1, route2, cost_eval), 0)
+
+    # Now we reverse the visits of route 1 and 2, so that we can hit the cases
+    # where route 1 is empty.
+    route1.clear()
+
+    for loc in [1, 2]:
+        route2.append(Node(loc=loc))
+
+    route1.update()  # depot -> depot
+    route2.update()  # depot -> 1 -> 2 -> depot
+
+    # The best move creates route1 as (depot -> 1 -> 2 -> depot) and route2 as
+    # (depot -> depot), making route 1 non-empty, while making route 2 empty.
+    # The total fixed cost incurred is thus 10 - 100 = -90.
+    assert_allclose(op.evaluate(route1, route2, cost_eval), -90)
+
+    op.apply(route1, route2)
+    assert_equal([node.client for node in route1], [1, 2])
+    assert_equal(len(route2), 0)
